feat(gc_safepoint): Skip test when NEXT_GEN is 1
- The `gc_safepoint` integration test is not compatible with the Next Generation GC.
- This commit adds a check to skip the test when `NEXT_GEN` environment variable is set to 1.

+++ tests/integration_tests/gc_safepoint/run.sh
@@ -74,6 +74,10 @@ export -f check_safepoint_equal
 export -f clear_gc_worker_safepoint
 
 function run() {
+	if [ "$NEXT_GEN" = 1 ]; then
+		exit 0
+	fi
+
 	rm -rf $WORK_DIR && mkdir -p $WORK_DIR
 	start_tidb_cluster --workdir $WORK_DIR
 	cd $WORK_DIR
